# Fast implementation of triangulation algorithms.