--- a/Project.toml
+++ b/Project.toml
@@ -33,7 +33,7 @@ StarAlgebras = "0c0c59c1-dc5f-42e9-9a8b-b5dc384a6cd1"
 
 [compat]
 AbstractPermutations = "0.3.1"
-ChordalGraph = "0.4.1"
+ChordalGraph = "0.4"
 CliqueTrees = "0.5.0"
 Cyclotomics = "0.3.2"
 DynamicPolynomials = "0.4.6"