--- a/src/TSSOS.jl
+++ b/src/TSSOS.jl
@@ -1,9 +1,9 @@
 module TSSOS
 
+using ChordalGraph
 using MosekTools
 using JuMP
 using Graphs
-using ChordalGraph
 using DynamicPolynomials
 using MultivariatePolynomials
 using Ipopt
@@ -19,12 +19,44 @@ using Random
 # using SDPT3
 # using SDPNAL
 
+import CliqueTrees
+
 export tssos_first, tssos_higher!, cs_tssos_first, cs_tssos_higher!, local_solution, refine_sol,
 cosmo_para, mosek_para, add_psatz!, add_poly!, get_nbasis, get_moment, get_moment_matrix, get_cmoment, homogenize, 
 solve_hpop, get_signsymmetry, SumOfRatios, SparseSumOfRatios, LinearPMI_first, LinearPMI_higher!, 
 show_blocks, complex_to_real, add_SOSMatrix!, sparseobj, get_mmoment, extract_solutions, extract_solutions_robust, 
 extract_solutions_pmo, extract_solutions_robust_pmo, extract_weight_matrix, add_SOS!, tssos_symmetry,
 run_H1, run_H1CS, run_H2, run_H2CS, construct_CDK, construct_marginal_CDK, construct_CDK_cs, construct_marginal_CDK_cs
+
+function chordal_cliques!(G; method="MF", minimize=false)
+    # choose algorithm
+    alg = method == "MF" && minimize  ? CliqueTrees.MinimalChordal(CliqueTrees.MF())  :
+          method == "MD" && minimize  ? CliqueTrees.MinimalChordal(CliqueTrees.MMD()) :
+          method == "MF" && !minimize ? CliqueTrees.MF()                              :
+          method == "MD" && !minimize ? CliqueTrees.MMD()                             :
+          error()
+
+    # compute maximal cliques
+    label, tree = CliqueTrees.cliquetree(G; alg)
+    
+    # triangulate graph
+    F = CliqueTrees.FilledGraph(tree)
+    
+    for edge in edges(F)
+        add_edge!(G, label[src(edge)], label[dst(edge)])
+    end
+    
+    # return maximal cliques
+    maximal_cliques = Vector{Vector{UInt16}}(undef, length(tree))
+    
+    for (i, clique) in enumerate(tree)
+        maximal_cliques[i] = sort!(label[clique])
+    end
+    
+    cliquesize = length.(maximal_cliques)
+    cql = length(cliquesize)
+    return maximal_cliques, cql, cliquesize
+end
 
 include("clique_merge.jl")
 include("blockpop_uncons.jl")