I have a package [CliqueTrees.jl](https://github.com/AlgebraicJulia/CliqueTrees.jl) with fast implementations of various graph triangulation algorithms. Here are some benchmarks.

```julia
julia> using Graphs, MatrixMarket, SuiteSparseMatrixCollection, SparseArrays, TSSOS

julia> ssmc = ssmc_db();

julia> name = "bcsstk11";

julia> matrix = mmread(joinpath(fetch_ssmc(ssmc[ssmc.name .== name, :], format="MM")[1], "$(name).mtx"));

julia> graph = Graph(fkeep!((i, j, x) -> i != j, matrix))
{1473, 16384} undirected simple Int64 graph
```

**old implementation**

```julia
julia> @time TSSOS.chordal_cliques!(copy(graph); method="MF", minimize=false);
  7.749093 seconds (96.07 k allocations: 67.437 MiB, 0.27% gc time)

julia> @time TSSOS.chordal_cliques!(copy(graph); method="MF", minimize=true);
  9.379268 seconds (311.25 k allocations: 142.738 MiB, 0.12% gc time)
```

**new implementation**

```julia
julia> @time TSSOS.chordal_cliques!(copy(graph); method="MF", minimize=false);
  0.011180 seconds (14.00 k allocations: 6.338 MiB)

julia> @time TSSOS.chordal_cliques!(copy(graph); method="MF", minimize=true);
  0.042636 seconds (22.98 k allocations: 11.396 MiB)
```