@samuelsonric 2025-03-04
Here, I am benchmarking the Broyden banded function from example 7.1 of [this paper](https://arxiv.org/abs/1912.08899).

```julia
julia> using DynamicPolynomials, TSSOS

julia> function broyden_banded(n)
        @polyvar x[1:n]
    
        g = sum(1:n) do i
            f = x[i] * (2 + 5x[i]^2) + 1
        
            for j in max(1, i - 5):min(n, i + 1)
                if i != j
                    f += (1 + x[j]) * x[j]
                end
            end
        
            return f^2
        end
    
        return g, x
    end;

julia> f, x = broyden_banded(500);
```

**old implementation**

```julia
julia> @time cs_tssos_first([f], x, 3; TS="MF");
 91.944828 seconds (5.83 G allocations: 219.252 GiB, 22.71% gc time)
```

**new implementation**

```julia
julia> @time cs_tssos_first([f], x, 3; TS="MF");
 78.821616 seconds (5.37 G allocations: 201.896 GiB, 23.82% gc time)
```